Fix global options swallowing arguments.
The --skip-root-check and --skip-config options were both defined as
InputOption::VALUE_OPTIONAL. In this situation the option will be
assigned the value of an argument that is directly following it. In the
case of a command this is the command name itself, but this causes every
argument to shift along. Since these options are boolean switches, they
can be changed to InputOption::VALUE_NONE. Closes #521.