[BUILD] require sudo on travis
fix missing suod requirements on travis for new builds.

when forking to a private repository and configuring it as new on travis,
the sudo requirement was not inherited as the project is new.

this commit requires sudo explicitly.

before ref: 9fb45b1 the problem of a missing sudo was unnoticed by the
build scripts on travis.