Fix install command
In a recent change 8115147 in 1.1.11 exceptions were swallowed by accident
leaving the install command with 0 exit status while being in error.

This qualifies as a flaw because the installer should exit early.

It came to light while getting Magento composer repository 500er errors
and then debugging the Magento install on Travis.

<https://repo.magento.com/packages.json> (with credentials)

    HTTP/1.1 500 Internal Server Error
    Cache-Control: no-cache
    Content-Type: application/json
    Date: Sat, 21 May 2016 07:30:17 GMT
    Server: nginx
    X-Frame-Options: DENY
    Content-Length: 57
    Connection: keep-alive

    {"error":{"code":500,"message":"Internal Server Error"}}

Likely the 500 error is a 503 given because the Magentocommerce website
login service that can be checked via <https://www.magentocommerce.com/
products/customer/account/login> is down.

    HTTP/1.1 503 Service Temporarily Unavailable
    Server: nginx
    Date: Sat, 21 May 2016 07:23:52 GMT
    Content-Type: text/html
    Content-Length: 70676
    Connection: keep-alive
    ETag: "573ff027-11414"

But this might only be a side-effect.

Additionally the -vvv switch has been added to the build to show error-
details for both the magerun and the composer command.

- Ref: https://travis-ci.org/ktomk/n98-magerun2/jobs/131873305#L422

- Ref: 8115147ddf726cc1c9203b98591f304853ebde0a

- Archive: http://archive.is/g4wBF