Check repository connectivity in Travis build
Script that allows to make Magento setup fail but not the entire build if
the reason is connectivity to the Mangeto composer repository.

This is a bit rough as the Magento repository is currently down but
builds are still needed.

Problems became immanent in ce5c3c5 and the verbose setting is now removed
again.

- Ref: ce5c3c5579e386fe0c622584347b628461619609