Fix install command replace-htaccess-file option, fixes #191
The code was a bit confusing, it could be clarified with the
following reference:

Do not ask any questions if "useDefaultConfigParams" is set, see
<https://github.com/netz98/n98-magerun/issues/480#issuecomment-72011750>.

then the fix was rather straight forward.

Additionally making the use of "useDefaultConfigParams" option as flag
according to command-line-help.

This has to be done as well for the install subcommand.