Fix install command use-default-config-params option
The "useDefaultConfigParams" option when set but with no value (in stand-
ard longopts style use, e.g. --useDefaultConfigParams) the option was not
interpreted as true/yes albeit set and the usage did list the value of it
as optional.

The fix is to interprete the flat as yes/true when set. This is aligned
with the flag usage in the rewrite-htaccess-file sub-command.

- Ref: 6d3b88