Fix deletion of complete config scope
Previously it was not possible to delete one or multiple paths from a
scope. It resulted in a fatal error as a wrong class was used for the
configuration data collection.

Fix is to use the correct model which needs injection.

Additionally made scope-id parameter truly optional so that it is possible
to actually detect if the scope-ids need to be obtained.