Auto-detect path for dev:urn-catalog:generate command
I was too lazy to type it all the time.

Warning: Implementation is more a POC than a stable feature. Symfony
console is not very user-friendly with command event and arguments.

One open idea is to fork another shell and call the command again
but with that parameter and cancel the original command instead. But that
is must tried first, too.