Check if it's possible to inject file, patches #233
Some other type of ArgvInput w/o that must be there, I could only
reproduce it once and it was too little to see what's happening there.

In case the file is found but the POC does not seem to be apllicable,
at least a hint to the filename is given to std-out.

Refs:

- https://github.com/netz98/n98-magerun2/issues/233

- 6abc8a74320b12bda5e7193264a6b52460a877de