Fix local test setup stopfile creation
when running from scratch the .n98-magerun stopfile was created in the
wrong folder.

This happened because within the ensure_magento2_auth function which is
called in the process the working directory is changed.

Creating the stopfile in the correct place and with changes of 81decd5
allows to additionally remove the hint on the environment variable
N98_MAGERUN2_TEST_MAGENTO_ROOT which is not needed any longer.

Refs:

- 81decd5eac4e7e2cd2014f8ef387f15068da15e1