Coding standards for code-generation
One step closer to make cs-checks part of the build.

Adding inline cs-fixing on generated code as the zend framework generators
aren't very sane regarding this and fixed files on disk for reference
require it's output a patch.

This is far from ideal, but does the job for a moment.

Also the situation is useful to deal with regarding test-case stream-
lining.