Fix removal of test-code
Test-code belongs to the repository itself, not into a library of it's own.

Fixes to the base test cases (two of them, a general one and a
specific one for developer console) will be applied to the files in
the code repository in a nearby, future commit.

It is important to have code and tests together in one repository. This
includes base testcases if they contain code specific helper methods
as we make use of them in the Magerun repository/ies.

This is a series of reverts in the following order:

Revert "[TASK] Updated test framework"

This reverts commit 2b8e30709ea0b57ae39e8537927b535fc2c85d20.

Revert "[TASK] Fix PHPCS messages (change order of uses)"

This reverts commit 38dace8d83f52a4f9cfa3878562214fc6486e577.

Revert "[TASK] Fix PHPCS messages"

This reverts commit 4e79f0ef745254f33b8ed0304f5446f75e288066.

Revert "[TASK] Add test framework"

This reverts commit 2d30606c690b3d49faca80414d1a329f5a376c24.