Fix root-dir behavior for magento root-dir detection
Use --root-dir value as first folder to detect magento in. Previously it
could be overridden by stopfiles. With this fix the root-dir folder is
added again as last folder which in reverse is then checked first as the
magento root directory.

Stopfiles still need to be resolved to find the stopfile directory as it
can hint further magerun extensions.