Build with PHP 7.1 for some jobs, tasks #256
Changes in preparation of use on system with PHP 7.1:

- Add PHP 7.1 lint check
- Add PHP 7.1 code-style fixer check
- Add PHP 7.1 phar build

Running the Magento 2 install with PHP 7.1 is not yet possible as it
fails. Officially Magento 2 is not yet compabitle with PHP 7.1 but
it also does not work due to the fact that Magento 2 uses the outdated
mcrypt library for "security" and we see deprecation notices in the
testsuite everywhere. Symfony has support for these migrations with
the PhpUnit Bundle which might be a place to look into.

Refs:

- https://github.com/netz98/n98-magerun2/issues/256