Add integration tester for Magerun commands
Writing quick integration tests is needlessly complex and created a
cultur of copy & pasta.

After a code duplication report from Scrutinizer we took the time to
review the existing test-code extracting the active parts out of it and
offer three simple assertion methods that are actually used:

- Command\TestCase::assertDisplayContains('com:mand', 'needle')

- Command\TestCase::assertDisplayRegExp('com:mand', '~pattern~')

- Command\TestCase::assertExecute('com:mand')

Alternatively instead of the command parameter the whole known Symfony
CommandTester $input parameter can be provided so that command name
and arguments can be passed.

The tester keeps automatically track so that multiple display assertions
execute the command / input a single time.

Additionally:

- Replace of Magento 1 test cruft with real Magerun 2 test code