Support setting config values to NULL, tasks #208
It was previously not possible to use config:set with NULL/unkown values
which are supported in the Magento config database store like so:

    config:set "web/unsecure/base_media_url"    NULL
    config:set "web/unsecure/base_static_url"   NULL

This commit adds support to set the null value unless --no-null is used
which will insert the string "NULL" which is the old behavior.

Additionally NULL values are properly reflected when displaying config data
via config:get making them more visible and even the concrete type in
JSON format.

Refs:

- #208

- Command: config:set

- Command: config:get