Fix bash autocompletion, fixes #331
Via `bin/compile-bash-autocompletion`.

The script

    build/bin/compile-bash-autocomplemetion-sh

creates (when invoked from a Zsh shell) an autocompletion file for Zsh.

This causes an error as the generated file is used in a bash environment.

Fix is to make the generation for the bash autocompletion explicit by
using the --shell=bash switch for the symfony-autocomplete tool.


Refs:

- #331

- #327