Pass DummyCommand to ConsoleEvent "before"
This solution is better than loading the wrong ConsoleEvent class. Old Magento 2.2.x installations have a incompatible ConsoleEvent class. The argument  "command" has to be a valid command. In Symfony 3.x the command argument can be null. At this moment we do not have a valid command, because we are before any command execution. So we pass a Dummy Command to the listener.