Prevent `Deprecated Functionality` under PHP 8,1 and default to all stores
While running the command `sys:url:list` under PHP 8.1 but not passing any `stores` as input argument, I got an error message `Deprecated Functionality: explode(): Passing null to parameter #2 ($string) of type string is deprecated in phar:///usr/local/bin/magerun2/src/N98/Magento/Command/System/Url/ListCommand.php on line 11`. The reason for this is that the call `$input->getArgument('stores')` returns `null` while `explode()` requires a string as second argument. This PR fixes that.

But while we're at it, the input argument `stores` is empty, so it makes more sense to default to a list of all stores. This PR adds that too.