We do not need to register Magerun as Magento "app"
We tried to fullfil the AppInterface of the Magento core all the
years. This is not needed to run the cli application.
In Magento 2.4.7 there are plugins on the AppInterface.
We do not want to trigger any code generation in the core application.
So I removed the compatbility layer which is not required.